% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rong.R
\docType{package}
\name{rong-package}
\alias{rong}
\alias{rong-package}
\title{rong: Functional Input Validation}
\description{
A set of basic tools to transform functions into functions with
input validation checks, in a manner suitable for both programmatic and
interactive use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/egnha/rong}
  \item Report bugs at \url{https://github.com/egnha/rong/issues}
}

}
\author{
\strong{Maintainer}: Eugene Ha \email{eha@posteo.de}

}
\keyword{internal}
